requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
